requirements.txt
cmake == 3.13.0
Flask==1.0.2
Jinja2==2.10
numpy==1.16.2
numpydoc==0.8.0
Pillow==6.2.1
torch==1.0.1
gunicorn == 19.9.0
torchvision==0.2.2
opencv-python-headless==4.0.0.21